requirements_complete.txt
# Smart Attendance System - Complete Requirements

# GUI Framework
kivy>=2.1.0
kivymd>=1.1.1

# Computer Vision & AI
opencv-python>=4.8.0
deepface>=0.0.79
tensorflow>=2.13.0  # Required by DeepFace
pandas>=2.0.0  # Required by DeepFace

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# Backend API (if using FastAPI)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4

# Additional dependencies
Pillow>=10.0.0  # Image processing
requests>=2.31.0  # HTTP requests (if needed)




